Update process data and get_screenshot functions for minimal dependencies
