+++ rigol_ds1054z/__init__.py
@@ -1,2 +1,1 @@
-from .rigol1000z import *  # noqa: F403
-from rigol_ds1054z.src.oscope import Rigol_DS1054Z as Rigol_DS1054Z
+from rigol_ds1054z.src.oscope import Oscilloscope as Oscilloscope
